+++ src/Akeneo/Category/back/tests/EndToEnd/InternalApi/UpdateAttributeControllerEndToEnd.php
@@ -150,7 +150,7 @@ class UpdateAttributeControllerEndToEnd extends ControllerIntegrationTestCase
         );
 
         $response = $this->client->getResponse();
-        $this->assertSame(Response::HTTP_NOT_FOUND, $response->getStatusCode());
+        $this->assertSame(Response::HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
     }
 
     public function testItThrowsErrorOnWrongType(): void
@@ -170,7 +170,7 @@ class UpdateAttributeControllerEndToEnd extends ControllerIntegrationTestCase
         );
 
         $response = $this->client->getResponse();
-        $this->assertSame(Response::HTTP_BAD_REQUEST, $response->getStatusCode());
+        $this->assertSame(Response::HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
     }
 
     public function testItDoesNotUpdateOnDeactivateTemplate(): void
@@ -215,7 +215,7 @@ class UpdateAttributeControllerEndToEnd extends ControllerIntegrationTestCase
         );
 
         $response = $this->client->getResponse();
-        $this->assertSame(Response::HTTP_NOT_FOUND, $response->getStatusCode());
+        $this->assertSame(Response::HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
     }
 
     private function createTemplate(): void