+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 - PIM-10958: Fix attribute option position after clicking on "done"
 - PIM-10976: Fix variant product counter on Product Model Edit Form for variant products without identifier
 - PIM-10983: Error HTTP 500 when adding a custom app
+- PIM-10980: Fix pagination update when applying filters on product association grid
 
 ## Improvements